test(cmd): consolidate run.go masking tests into run_masking_test.go
Merge logoutput_test.go and collect_run_secrets_test.go (both testing
run.go functions) into one file; keep logger_test.go separate as it pairs
with logger.go.
